misc: new driver to set device specific parameters for GPU
Introducing a new driver to set device specific parameters
for GPU. With this driver we can set the driver data and
dma coherent mask for vfio kgsl device, by binding this driver
with the device first. Once the device specific configs are
set, we can bind the device with the vfio-platform driver to
probe the initialize the device as a vfio device.

Signed-off-by: Aravind Vijayakumar <quic_aprasann@quicinc.com>
Change-Id: Ic3bd217e44ff70b09a3a9b5339160f81b5dd2ced