kiumd: make the module compatible with debug build
For debug build, with lot of debug configs enabled,
the physical address of the buffer is mangled during
mapping, and there is corresponding unmangling during
unmap. So iommu_map_sg is expecting the mangled sg table,
in regular map using dma_buf_map_attachment, the kernel
takes care of the mangling, but incase of iommu_map_sg
kiumd have to take care of mangling before invoking the
funtion.

set max segment size and use the unlocked variant of
dma_buf_map_attachment. Mangle the dma-buf for debug build.
Mangling is only in debug build so enable the mangling
only for debug build.

Change-Id: I3bed2b87d61fd6e8888e7b67168034d8b77d6dc2
Signed-off-by: Aravind Vijayakumar <quic_aprasann@quicinc.com>