kiumd: introduce new iova allocator for dma-buf map
The existing kernel iova allocator allocates iova and aligns it
to size of the buffer(when size is power of 2), so if the size
of buffer is large like greater than 2 GB, the iova allocation
fails if the address space is limited to 32 bit. Multiple DSPs
have the limitation of 32 bit iova space.

so to resolve that, introduce a new iova allocator in kiumd
which aligns the iova to buffer size upto a limit and for buffers
greater than this limit is aligned to the max_shift which is
obtained from the device tree parameter "qcom,iova-align-shift-max"

Change-Id: I8848fd6e6d8c99735323c27c96a07474ff57dff4
Signed-off-by: Aravind Vijayakumar <quic_aprasann@quicinc.com>
(cherry picked from commit 99aea0b6f616e3b483c32e1a93349d30de1275be)