kiumd: fix for successive vfio device initialization failure
In case of relaunch of UMD applications they are calling vfio
init api again to initalize vfio leading to fauilure while
vfio group open system call resulting in application failure.
Root cause of this issue is while dmabuffer unmapping we are
getting the reference of vfio device via fget but we are closing
the file in other block which should be closed in outer block.
To fix this placed the file clousure to end of unmap api.

Change-Id: I57c58ce51208457a5eaf9fed9835a49d02b6a968
Signed-off-by: Vikas Srivastava <quic_vikassri@quicinc.com>