kiumd: invoke runtime suspend for vfio device after init
Invoke runtime suspend for vfio device after vfio device
initialization. vfio device initialization invokes runtime
resume for devices suppliers which include SMMU. For GPU,
this is holding the vote to HLOS CX gdsc, because of that
cx head switch is on, cx head switch needs to be turned off for
GPU hard reset.

Change-Id: I8e73416decb34a5a8a8639a1b9105f84727e791e
Signed-off-by: Aravind <quic_aprasann@quicinc.com>