drivers: Introduce Qualcomm Functional Safety (FuSa) framework support for TLMM
TLMM FuSa interrupts can be triggered due to 2D CSR parity errors.

The driver registers the handler for TLMM IRQ - tlmm_csr_aggregated_irq,
and monitors for any type of 2D CSR parity FUSA interrupts.

There are 2 types of such errors:
1.) AHB_HRESET
This stands for Advanced High-performance (AHB) Hardware Reset Error
This error occurs when there is a reset signal on the AHB interface
2.) POR_ARES
This stands for Power-on-reset (POR) Asynchronous Reset Error (ARES)
This error occurs when there is a power-on reset signal error

On the occurrence of the interrupts, the driver does the following:
Detects the type of error -> Acknowledges the interrupt -> Logs the errors

TLMM FuSa driver handles the registration and handling of all corresponding H/W IRQs introduced in the TLMM FuSA block.
IRQ handler for these faults, is tasked to read the interrupt status register and then clear the corresponding interrupt.

Below mechanism are made available to convey this information to the
userspace,
	a. TraceFS (via trace event)
	b. DebugFS (via dedicated file per TLMM instance)

Change-Id: I26a239d1129527002adc2c546d547485d10c2762
Signed-off-by: Sankalp Negi <quic_snegi@quicinc.com>