kiumd: Fix iova allocation reset when address space is exhausted
When the iova space is exhausted, the allocation short restart
from start_iova. However, the function was missing the reset
to start_iova, causing incorrect behavior when searching for
holes.

This fix ensures that last_allocated_end is reset to start_iova
when the allocations reaches the end of the iova range.

Fixes: a581659a2723b3ac27ec4e779b643bef2590215f
Change-Id: I96bc66de3189aba497e3dae4252d7c19699d77a0
Signed-off-by: Aravind Vijayakumar <aprasann@quicinc.com>