kiumd: Set the pagetable ops only if TTBR1 is programmed
During KGSL initialization, TTBR0 is configured by default.
However, LPAC requires TTBR1 to be programmed instead.
This patch ensures that TTBR1 is initialized only once
during the initial KGSL launch.

The TTBR1 setup is triggered from the VFIO device initialization.
To prevent unnecessary reconfiguration, the code checks if TTBR1
has already been programmed and skips reinitialization if so.

Also, the page table ops are set only if TTBR1 is successfully
programmed. If not, a warning is logged and the setup is aborted
gracefully.

Change-Id: I3c7859840421addce0b4d857a0c4df18c949aeb7
Signed-off-by: Aravind Vijayakumar <quic_aprasann@quicinc.com>