safelinux-modules: use wildcards in spec file
Previously the RPM spec file listed each kernel module and header file
explicitly in the %files lists. This would require changes every time a
new file is installed or a file is no longer installed.

Now the %files lists use wildcards to package all header files and
kernel modules. This will no longer require spec file changes when the
header file or module lists change. The downside is that the RPM build
will not error when we accidentally change the set of kernel modules or
header files being built.

Change-Id: I072afa6e1e789a4c83b5d033e3e847c1feea8de6
Signed-off-by: Eric Smith <quic_ericsmit@quicinc.com>