kiumd: remove unnecessary NULL pointer checks at start of IOCTL
Kiumd IOCTLs are checking for kiumd_ctx and fp at the
start of each IOCTL. This is redundant because:
a) IOCTLs are stopped at the VFS layer if there is no file
backing the fd associated with the driver.
b) kiumd_ctx is private data to the driver and is only freed
in the kiumd release method (called kiumd_close), which can
only be invoked when all IOCTLs are closed.

Removed these redundant checks to improve code quality.

Change-Id: Ib419bca72b4ee59ea11f730c92f22d1194500da5
Signed-off-by: Archit Saxena <quic_archsaxe@quicinc.com>