kiumd: Update struct smmu_map_data fields
Current implementation of smmu_map_data has fields marked long as they
used to be returned to userspace. This is not the case currently. This
makes unmap functions call unnecessary typecasts.

Mark all fields in smmu_map_data to the actual type they represent in
place of generic long type.

Change-Id: I1e206a5a7df312b72276d03f602381e6cc0c69e6
Signed-off-by: Archit Saxena <archsaxe@qti.qualcomm.com>