drivers: Introduce fault injection support for SMMU FuSa
Currently Qualcomm Technologies, Inc. automotive platforms support
two types of SMMU Client relation:
	1. TCU-TBU500 (Ex. SA8225, SA7255)
	2. TCU-QTB500 (Ex. SA8797)

For both cases, an additional hardware functionality is added in the
Qualcomm SMMU sub-system, directly related to automotive safety,
called as Functional Safety (FuSa) block. The support for enabling
SMMU FuSa block was added in this change:
commit 492ff9f35813 ("drivers: Introduce Qualcomm Functional Safety
(FuSa) framework support for SMMUv2.")

We now introduce a mechanism to validate this H/W FuSa block before
we resume normal platform execution. To do so, we inject each
applicable SMMU FuSa interrupt, and then monitor the corresponding
interrupt status register to ensure that the block functions correctly
as expected. Failure of  this injection test signifies a compromise
in the system safety, which could result in an undefined scenario.
So, as a response to this test failure, a system panic is introduced
to halt progression of normal platform execution.

SMMUv2 FuSa interrupts injected for each available TCUs:
	-> TCU SRAM parity
	-> TBU SRAM parity
        -> CSR parity
        -> PTW 2D Array parity

On the client side, error injection is possible only of platform with
TCU-TBUs support. In this case, following interrupt are injected for
each TCU-TBU pair:
	-> TBU VA error
	-> TBU PA error
	-> TBU WR Buffer Error
	-> TBU WR Buffer Warning
	-> TBU-TCU link parity

Change-Id: I498500f8990df0874023db12037182004f79f0d5
Signed-off-by: Ninad Naik <ninanaik@qti.qualcomm.com>